+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "jackiedo/log-reader",
-    "description": "An easy log reader and management tool for Laravel 4",
+    "description": "An easy log reader and management tool for Laravel",
     "keywords": ["reader", "log reader", "log viewer", "log manager", "log", "logs", "laravel"],
     "license": "MIT",
     "authors": [
@@ -10,14 +10,11 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "illuminate/support": "4.2.*"
+        "php": ">=5.6.4",
+        "illuminate/support": "5.3.*"
     },
     "autoload": {
-        "classmap": [
-            "src/migrations"
-        ],
-        "psr-0": {
+        "psr-4": {
             "Jackiedo\\LogReader\\": "src/"
         }
     },
