+++ src/Contracts/Patternable.php
