+++ src/Entities/LogEntry.php
