--- src/Jackiedo/LogReader/LogReaderServiceProvider.php
@@ -1,47 +1,0 @@
-<?php namespace Jackiedo\LogReader;
-
-use Illuminate\Support\ServiceProvider;
-
-class LogReaderServiceProvider extends ServiceProvider {
-
-	/**
-	 * Indicates if loading of the provider is deferred.
-	 *
-	 * @var bool
-	 */
-	protected $defer = false;
-
-	/**
-	 * Bootstrap the application events.
-	 *
-	 * @return void
-	 */
-	public function boot()
-	{
-		$this->package('jackiedo/log-reader');
-	}
-
-	/**
-	 * Register the service provider.
-	 *
-	 * @return void
-	 */
-	public function register()
-	{
-		$this->app['log-reader'] = $this->app->share(function($app)
-		{
-			return new LogReader;
-		});
-	}
-
-	/**
-	 * Get the services provided by the provider.
-	 *
-	 * @return array
-	 */
-	public function provides()
-	{
-		return ['log-reader'];
-	}
-
-}
